test_modules/netperf: fix non omni output parsing
There is a bug when running non omni output test, like SCTP tests,
with nperf_cpu_util alias not defined, the value stored in
self._cpu_util is None and not 'None' and the result rate is set 0
because the condition is met, bad regex is created and this regex
does not match the output.

This patch fixes that.